backend/requirements.txt
asgiref==3.8.1
cassandra-driver==3.29.2
click==8.1.7
confluent-kafka==2.6.0
cqlsh==6.2.0
Django==5.1.3
dnspython==2.7.0
eventlet==0.37.0
geomet==0.2.1.post1
gevent==24.11.1
greenlet==3.1.1
kafka-python==2.0.2
pure-sasl==0.6.2
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.1
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.1.1
